Move solver->gecode_solver in preparation for ruby_solver
