+++ lib/solve/gecode_solver.rb
