+++ spec/unit/solve/gecode_solver_spec.rb
